python-app/requirements.txt
python-jose==3.3.0
pyjwt==2.0.0